factored out stuff from ann package
