+++ src/test/java/io/anserini/analysis/FeatureVectorTokenizerTest.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package io.anserini.ann;
+package io.anserini.analysis;
 
+import io.anserini.analysis.FeatureVectorsTokenizer;
 import org.apache.lucene.analysis.Tokenizer;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
 import org.junit.Test;
