+++ src/test/java/io/anserini/analysis/fw/FakeWordsEncoderAnalyzerTest.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package io.anserini.ann.fw;
+package io.anserini.analysis.fw;
 
 import io.anserini.analysis.AnalyzerUtils;
+import io.anserini.analysis.fw.FakeWordsEncoderAnalyzer;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
