+++ src/test/java/io/anserini/analysis/lexlsh/LexicalLshAnalyzerTest.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package io.anserini.ann.lexlsh;
+package io.anserini.analysis.lexlsh;
 
 import io.anserini.analysis.AnalyzerUtils;
+import io.anserini.analysis.lexlsh.LexicalLshAnalyzer;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.TextField;
