+++ src/test/java/io/anserini/analysis/lexlsh/LexicalLshFeaturePositionTokenFilterTest.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package io.anserini.ann.fw;
+package io.anserini.analysis.lexlsh;
 
-import io.anserini.ann.FeatureVectorsTokenizer;
+import io.anserini.analysis.FeatureVectorsTokenizer;
+import io.anserini.analysis.lexlsh.LexicalLshFeaturePositionTokenFilter;
 import org.apache.lucene.analysis.Tokenizer;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
 import org.junit.Test;
@@ -28,44 +29,44 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 
 /**
- * Tests for {@link FakeWordsEncodeAndQuantizeFilter}
+ * Tests for {@link LexicalLshFeaturePositionTokenFilter}
  */
-public class FakeWordsEncodeAndQuantizeFilterTest {
+public class LexicalLshFeaturePositionTokenFilterTest {
 
   @Test
   public void testFiltering() throws Exception {
     StringReader reader = new StringReader("-0.10 0.20 0.30 0.40");
     Tokenizer stream = new FeatureVectorsTokenizer();
     stream.setReader(reader);
-    FakeWordsEncodeAndQuantizeFilter filter = new FakeWordsEncodeAndQuantizeFilter(stream, 20);
+    LexicalLshFeaturePositionTokenFilter filter = new LexicalLshFeaturePositionTokenFilter(stream);
     filter.reset();
     List<String> expectedTokens = new LinkedList<>();
-    expectedTokens.add("_"); // quantization leads to zero
-    expectedTokens.add("f2"); // quantization leads to 4 tokens
-    expectedTokens.add("f2");
-    expectedTokens.add("f2");
-    expectedTokens.add("f2");
-    expectedTokens.add("f3"); // quantization leads to 6 tokens
-    expectedTokens.add("f3");
-    expectedTokens.add("f3");
-    expectedTokens.add("f3");
-    expectedTokens.add("f3");
-    expectedTokens.add("f3");
-    expectedTokens.add("f4"); // quantization leads to 16 tokens
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
-    expectedTokens.add("f4");
+    expectedTokens.add("1_-0.10");
+    expectedTokens.add("2_0.20");
+    expectedTokens.add("3_0.30");
+    expectedTokens.add("4_0.40");
+    int i = 0;
+    while (filter.incrementToken()) {
+      CharTermAttribute charTermAttribute = filter.getAttribute(CharTermAttribute.class);
+      String token = new String(charTermAttribute.buffer(), 0, charTermAttribute.length());
+      assertEquals(expectedTokens.get(i), token);
+      i++;
+    }
+    filter.close();
+  }
+
+  @Test
+  public void testFilteringWithOffset() throws Exception {
+    StringReader reader = new StringReader("-0.104123 0.20435 0.3042366 0.41243241");
+    Tokenizer stream = new FeatureVectorsTokenizer();
+    stream.setReader(reader);
+    LexicalLshFeaturePositionTokenFilter filter = new LexicalLshFeaturePositionTokenFilter(stream, 2);
+    filter.reset();
+    List<String> expectedTokens = new LinkedList<>();
+    expectedTokens.add("1_-104123");
+    expectedTokens.add("2_20435");
+    expectedTokens.add("3_3042366");
+    expectedTokens.add("4_41243241");
     int i = 0;
     while (filter.incrementToken()) {
       CharTermAttribute charTermAttribute = filter.getAttribute(CharTermAttribute.class);
